Changing return value of __hash__ to size_t
SWIG is having a hard time coping with uintptr_t

+++ casadi/core/matrix/matrix_impl.hpp
@@ -1682,7 +1682,7 @@ namespace casadi {
   }
 
   template<typename DataType>
-  long Matrix<DataType>::getElementHash() const {
+  size_t Matrix<DataType>::getElementHash() const {
     throw CasadiException("\"getElementHash\" not defined for instantiation");
   }
